Fix namespaces and reorganize file structure
